assistent/requirements.txt
# tested on Ubuntu 20.04
# use repositoryies and packages:
# whisper:  https://github.com/openai/whisper/tree/main
# sounddev: https://readthedocs.org/projects/python-sounddevice/downloads/pdf/latest/
# miniconda:https://docs.conda.io/en/latest/miniconda.html
# Llama2
# obabooga
# bark:     https://huggingface.co/suno/bark
#           sudo apt update && sudo apt install ffmpeg#git+https://github.com/openai/whisper.git
#git+https://huggingface.co/suno/barkTTS
bark_ssg==1.3.4
pynput==1.8.1
pyperclip==1.11.0
Requests==2.32.5
scipy==1.17.0
sounddevice==0.5.2
soundfile==0.13.1
openai-whisper
